Localize the description
So I thaught that it was ackward that the description was in
english even when the calendar was selected in german. Now
we detect the language of the calendar and generate the description
accordingly. ^^

There is still some problems, like all Room information is only
available in english. But that is still progress.