Improve shorthands for english calendars
Till now shorthands were only detected on german calendars, but
ofcourse if the calendar can be exported in english we should
support that. So I went through the tedious process of finding
all the english lecture names and adding them to the csv.

This now means we totally support english and german ane the
users can choose whatever they are most comfortable with. ^^