Add tests and restructure project
In order to add tests I had to restructure the project to use
a package for the flask app. This makes is a little easier to
find the backend code but overall I think that many other languages
force better layouts.

Anyway, I am actually quite excited to add some tests as the app
was getting large enough to have some anxiety that stuff would
break when adding features. Since students should be able to rely
on this project I really want to have the confidence that nothing
randomly breaks.

I still have to read more into mocking and working with pytest
but so far I am quite excited. ^^