Add tests against dropped events and left template strings
In one of my previous commits I had a template string in the
description because I forgot to put an f into the f-string. Now
there are tests that make sure this never happens again. It is now
also impossible to drop events in the mapping process as the tests
verify that the new calendar has as many events as the old one.