Add room generator script and update room list
Thanks to @bananensplit I now have the original script that generated
rooms.csv. I updated it a bit but still most of the credits go out to
him.

With that script I went ahead and updated the list of rooms which
changed quite a bit. So happy that this is now automated. ^^