Add fallback shorthands #2
The results are suprisingly good, I know that hand-picked
shorthands are better but it would be impossible to generate them
for all lectures at the university also because its a social thing.

This however should be the best middleground and if the generated
shorthands aren't good enough we can still add custom ones to the
data list.