Fix the floor fix
So I previously was right that the floor parsing could have
provided wrong floor information. However, my fix also was not
quite right as the field in which the floor information is can
change. This fix should be much better at finding the right field
but in general the dataset is just not that good.

Maybe we can do better at generating the dataset. This mainly
depends on how clean the data from TISS is but I will look into
that.